Update 5_generate_TL_file.py
Make it so the script won't stop if it runs into an "avgID" for a file that isn't present in the "MonoBehaviour" folder (like a custom renpy script or something).
This may also make it possible for an actual avg file to be unintentionally skipped, so added an error message at the end if a file was skipped.